Add SSH support for containers
-Installs openssh-server
-Imports public keys via `ssh-import-id` using GitHub username and
adds them to '/root/.ssh/authorized_keys' if GH_USER is set

This is 'secure' in that use of public keys are always secure, but it is
not the wisest strategy to use the same key-pair for multiple venues
(server cluster, public website, etc.). More robust key management is a
feature for a future date. For now, this suffices for protypting and
development on small clusters.