Force supervisor.conf symlink creation.
This is needed in situations where we use the `--net host` option in
docker.  The unique ID for $HOSTNAME becomes the host's hostname in this
case and symlink collisions are going to occur when the hub is
persistent, but the spoke is starting up for the first time. The old
symlink and $HOSTNAME folders will already exist, so we make all the
creation steps involving $HOSTNAME happen regardless.