Add override to let spokes run without hub.
SPOKE_DETACH_MODE=False by default. Because not all Spokes HAVE to run
with hubs, even though they probably will if they are a typical "stack"
type of application. Some programs truly are stand alone binaries
without the need for configuration OR just have a need for extreme
portability. The option to have a stone-alone mode could be useful on a
Spoke-by-spoke basis.